fix json part
